Simplify auth scheme
Relax the "credentials required" restriction and instead just use the
middleware to obtain and validate any submitted authentication
credentials. Any particular operation that requires permissions is going
to care not just that the user is logged in, but that their particular
account has some specific permission (as in the superuser case). So the
previous "require auth" layer is never going to be exactly what we need.