Add some indexing for timeline sorts
Azure's (re-)implementation of MongoDB requires that fields that you
sort on are indexed, which actual Mongo doesn't. But that's pretty
reasonable! Add the indices in our DB setup, and change the "home"
timeline to sort by creation_date since I don't want to set up an index
on `text` just for that random choice.