Correct CORS origin header
We used to return `Access-Control-Allow-Origin: *` as a header. But for
the session cookie, that's not allowed. I then tried to set up the CORS
module with a potentially array of allowed origins, but that wasn't
correct either. It is OK for requests to come in from unexpected
origins (like the previewer server, which revealed this problem);
it's just that we can't indicate this using the `*` syntax.

This is exactly what the `origin: true` setting does -- it allows
requests but "reflects" back the provided origin, so that the client
always sees itself as being approved.

For something like our session-cookie API, I was initially thinking that
it might be better to restrict the origin — ideally you could only
create a session through the official frontend. But there is nothing
preventing someone from faking the `Origin:` header, so any kind of
restriction isn't going to add any actual security. As long as we're
accepting requests from any IP address, I don't see how origin filtering
can help with anything.