Try converting to ESM modules instead of CJS
The @wwtelescope/astro module is distributed as ESM, and ESM only, and
it seems like the easiest approach is to use ESM here too to keep things
compatible. At least, it seems easier to change here rather than to
update the wwt-webgl-engine build systems in a way that could be a bit
fragile/risky.

This is largely straightforward, but we need to update a bunch of our
import statements to specify the extension of the imported file. As per
some extensive discussions on the TypeScript GitHub, the way to do this
is to use `.js` even though we're writing TypeScript. Also, it seems
that a few modules like `fp-ts` need to be imported via their `/lib/`
subdirectories rather than `/es6/` even though I'd have thought that the
latter would be preferable.