algorithm.ts: fix behavior when an initial scene is specified
We were failing to remove the initial scene from the list of candidates,
so its ordering was getting overwritten later during the bulk
processing.

The basic issue is that Mongo ObjectIds are JS objects and so in the
general case:

```
new ObjectId("abc123") != new ObjectId("abc123")
```

To test equality, we instead have to use `ObjectId.equals()`.

Also fix up shadowing of the `handles` variable.